feat: implement basic commands logic
Currently, the persistence is not implemented.

Also, the ID has been downgraded from `u64` to `u32`

+++ src/cli.rs
@@ -25,11 +25,11 @@ pub enum Command {
     },
 
     Done {
-        id: u64,
+        id: u32,
     },
 
     Remove {
-        id: u64,
+        id: u32,
     },
 }
